Add check option and load check modules
This commit adds the ability to load and run check modules.
The check modules can be added as children of Check.pm. The check
modules will inherit Check.pm's methods and constants for
running commands and returning data.

Instructions and examples for expected methods and return data
structure have been added to the documentation of the parent
module, Check.pm.