Mock get_avail_zone to avoid meta-data queries
CloudWatchClient::call_json gathers various bit of instance
information from the host through meta-data calls to the
filesystem.  If tests are run on a machine that's not an AWS
instance, the meta-data calls will fail and ultimately hit the 2
second timeout.  Each failed meta-data call adds 2 seconds to
the runtime of the test.

This commit additionally mocks out the get_avail_zone meta-data
call within monitor-run.t to speed up the test and remove
unnecessary calls to resources that don't exist on development
hosts.