Fix undef warning on chomp in read_meta_data
Every time read_meta_data reads a metadata value, regardless if
it's read from the cache or retrieved fresh from the special
AWS mount, read_meta_data is re-writing the value back into the
cache file.

In situations where more than one aws-cloudwatch-monitor process
is running, for example two cronjobs each running at the top of
the minute, read_meta_data's frequent cache writing causes a
race condition/collision where the value isn't written to the
file as the other process reads it.

If the code reads an empty value (in the case of the race
condition), it will get a new value fresh from the instance (we'll
never hit a situation where we don't have a value).

More work needs to be done to improve the caching read/write so
we're not writing the value again after every read (which is an
inefficient cache design).  For now, this commit fixes the warning
by not attempting to chomp a potentially undef value.