Fix process check math for zombie processes
If the process being checked is marked as zombie (defunct), the
ps output returns the <defunct> string in the second column where
the pcpu column is expected to be.

When the check module attempts to do math on a string, perl
produces an "isn't numeric in addition" warning since the
string doesn't "look like a number" to Perl.

https://perldoc.perl.org/perlapi#looks_like_number

This commit fixes the process check to ignore zombie processes by:

- adding the stat column to the ps output
- checking the stat column to see if the process is marked zombie
- incrementing values only for processes not marked zombie