Fix meta-data caching writes
The meta-data cache files are written after every call, regardless
of the TTL variables.  This presents a couple issues.

First, the age of the cache files never get older than the time of
the last call, effectively making the TTL time meaningless.  This
leads to far too many unnecessary writes to the disk since the
cache files never age.

Second, since the TTL time is not honored, if the special meta-data
mount is incorrectly returning empty string (an issue at AWS or the
mount on the instance), the benefit of relying on a cached value is
lost.

This commit fixes the meta-data cache writes to now only write the
cache file if a value was retrieved from the special mount, which
only happens if the TTL is reached or the cache file doesn't exist
or doesn't contain a value.

Additionally, if the TTL of the cache has expired and the special
mount didn't return data, the expired value from the cache will now
continue to be used until the mount starts to return data again.

New warnings have also been added to inform the admin that expired
data is being used, and that no data was found in either the cache
or mount.

This commit also updates the read_meta_data call within
get_auto_scaling_group to expect the new return and to check
ttl_expired before fetching the new data from the API.